Move remove functionality to its own module
